+++ lib/molecules/toast/zwapToast.dart
@@ -10,7 +10,7 @@ import 'package:zwap_design_system/atoms/atoms.dart';
 /// Custom class to show toast message
 class ZwapToasts {
   static Widget toastWidget(String normalText, bool isError, double width, Function() dismiss) {
-    Size size = getTextSize(normalText, ZwapTextType.h3);
+    Size size = getTextSize(normalText, ZwapTextType.mediumBodyMedium);
     int line = (size.width / (width - 130)).floor();
 
     return Align(
@@ -19,21 +19,22 @@ class ZwapToasts {
         borderRadius: BorderRadius.circular(10.0),
         child: Container(
           decoration: BoxDecoration(
-              color: isError ? ZwapColors.error400 : ZwapColors.success700, boxShadow: [BoxShadow(color: Color.fromRGBO(199, 199, 199, 0.5), blurRadius: 30, spreadRadius: 4, offset: Offset(0, 15))]),
+              color: isError ? ZwapColors.error400 : ZwapColors.success400,
+              boxShadow: [BoxShadow(color: Color.fromRGBO(199, 199, 199, 0.5), blurRadius: 30, spreadRadius: 4, offset: Offset(0, 15))]),
           width: width,
           height: 48 + max(line * 26, 0),
           padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 20),
           child: Row(
             crossAxisAlignment: CrossAxisAlignment.center,
             children: <Widget>[
-              ZwapText(text: isError ? '😵' : '✅', zwapTextType: ZwapTextType.bodySemiBold, textColor: ZwapColors.shades0),
+              ZwapText(text: isError ? '😵' : '✅', zwapTextType: ZwapTextType.mediumBodyMedium, textColor: ZwapColors.shades0),
               SizedBox(width: 20),
-              Expanded(
+              Flexible(
                 child: ZwapText(
                   text: normalText,
                   textColor: Colors.white,
                   textAlign: TextAlign.center,
-                  zwapTextType: ZwapTextType.h3,
+                  zwapTextType: ZwapTextType.mediumBodyMedium,
                 ),
               ),
               SizedBox(width: 20),